+++ forui/test/src/widgets/picker/time/time_picker_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
 import 'package:forui/forui.dart';
+import 'package:forui/src/widgets/picker/picker_controller.dart';
 import 'package:forui/src/widgets/picker/picker_wheel.dart';
 import 'package:forui/src/widgets/picker/time/time_picker_controller.dart';
 import '../../../test_scaffold.dart';